Make JXTypeID.{name,protocols} settable
This is useful if, for example, we need to demangle Swift names, since we can simply modify the type in-place

+++ TypeParser/src/JXTypeArray.m
@@ -26,7 +26,7 @@
 
     unsigned long long count;
     if (![scanner scanUnsignedLongLong:&count]) return nil;
-    _count = count;
+    _count = (NSUInteger)count;
 
     _type = [JXType typeWithScanner:scanner];
